[FEAT]생각 작성 후 return으로 데이터 전부 가져오도록 수정

+++ src/main/java/com/async/main/domain/cleaningGauge/entity/CleaningGauge.java
@@ -29,6 +29,6 @@ public class CleaningGauge extends BaseTimeEntity {
     private String componentsType;
 
     @Column(name = "clean_gauge")
-    private Double cleanGauge;
+    private Float cleanGauge;
 }
